refactor folder structure, change post.js to be reusable
