+++ src/components/Home/index.js
@@ -1,11 +1,11 @@
 import React, { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import Banner from "../components/Banner/Banner";
-import Layout from "../components/common/Layout";
-import LoadMoreBtn from "../components/Home/LoadMoreBtn";
-import PostList from "../components/Home/Post/PostList";
-import PostListSkeleton from "../components/Home/Post/PostListSkeleton";
-import { getPostsAction } from "../store/reducers/postsReducer";
+import Banner from "./../Banner/Banner";
+import Layout from "../common/Layout";
+import LoadMoreBtn from "./LoadMoreBtn";
+import PostList from "./Post/PostList";
+import PostListSkeleton from "./Post/PostListSkeleton";
+import { getPostsAction } from "../../store/reducers/postsReducer";
 export default function Home() {
   const dispatch = useDispatch();
   const { data, loading, error } = useSelector((state) => state.posts);
