+++ src/components/Login/index.js
@@ -1,20 +1,21 @@
-import { Button, Card, Form } from "@ahaui/react";
-import React, { useEffect } from "react";
-import styled from "styled-components";
-import Layout from "../components/common/Layout";
-import { Navigate } from "react-router-dom";
-import { useAuth } from "../hooks/useAuth";
-import { useForm } from "react-hook-form";
-import {
-  registerAction,
-  resetErrorAction,
-} from "../store/reducers/authReducer";
+import { Button, Card, Form, Loader } from "@ahaui/react";
+import React from "react";
 import { useDispatch } from "react-redux";
-import { AUTH_ERROR_MESSAGES } from "../constants/Auth/Message";
-export default function Register() {
-  const { isLoggedIn } = useAuth();
-  const dispatch = useDispatch();
+import { Navigate } from "react-router-dom";
+import styled from "styled-components";
+import Layout from "../common/Layout";
+import { useAuth } from "../../hooks/useAuth";
+import {
+  loginAction,
+  resetErrorAction,
+} from "../../store/reducers/authReducer";
+import { useForm } from "react-hook-form";
+import { AUTH_ERROR_MESSAGES } from "../../constants/Auth/Message";
 
+export default function Login() {
+  const { isLoggedIn, loading, error: errorApi } = useAuth();
+  const dispatch = useDispatch();
+  console.log({ errorApi });
   const {
     register,
     handleSubmit,
@@ -22,9 +23,8 @@ export default function Register() {
     formState: { errors },
   } = useForm();
   const onSubmit = ({ username, password }) => {
-    console.log(`onSubmit`, { username, password });
     dispatch(
-      registerAction({
+      loginAction({
         username,
         password,
       })
@@ -35,25 +35,20 @@ export default function Register() {
     return <Navigate to="/" />;
   }
 
+  const isHasUsernameError = errorApi?.username || errors?.username;
+  const isHasPasswordError = errors?.password || errorApi?.password;
   return (
     <Layout>
-      <RegisterWrapper>
+      <LoginWrapper>
         <Card style={{ height: "fit-content" }} size={"medium"}>
           <Card.Body>
             <form onSubmit={handleSubmit(onSubmit)}>
               <div>
-                <Form.Group controlId="registerForm.email">
-                  <Form.Label>Email</Form.Label>
-                  <Form.Input
-                    type="text"
-                    placeholder="Enter email"
-                    defaultValue=""
-                  />
-                </Form.Group>
-                <Form.Group controlId="registerForm.username">
+                <Form.Group controlId="loginForm.username">
                   <Form.Label>Username</Form.Label>
                   <Form.Input
                     type="text"
+                    isInvalid={isHasUsernameError}
                     placeholder="Enter text"
                     {...register("username", {
                       required: AUTH_ERROR_MESSAGES.USERNAME_REQUIRED,
@@ -62,11 +57,17 @@ export default function Register() {
                       },
                     })}
                   />
+                  {isHasUsernameError && (
+                    <Form.Feedback type="invalid">
+                      {errorApi?.username || errors?.username.message}
+                    </Form.Feedback>
+                  )}
                 </Form.Group>
-                <Form.Group controlId="registerForm.password">
+                <Form.Group controlId="loginForm.password">
                   <Form.Label>Password</Form.Label>
                   <Form.Input
                     type="password"
+                    isInvalid={isHasPasswordError}
                     placeholder="Enter password"
                     {...register("password", {
                       required: AUTH_ERROR_MESSAGES.PASSWORD_REQUIRED,
@@ -75,24 +76,32 @@ export default function Register() {
                       },
                     })}
                   />
+                  {isHasPasswordError && (
+                    <Form.Feedback type="invalid">
+                      {errorApi?.password || errors?.password.message}
+                    </Form.Feedback>
+                  )}
                 </Form.Group>
                 <Button
                   size={"small"}
+                  type="submit"
                   variant="primary"
                   className="u-marginRightSmall"
                 >
-                  <Button.Label>Register</Button.Label>
+                  <Button.Label>
+                    {loading ? <Loader size="small" /> : "Login"}
+                  </Button.Label>
                 </Button>
               </div>
             </form>
           </Card.Body>
         </Card>
-      </RegisterWrapper>
+      </LoginWrapper>
     </Layout>
   );
 }
 
-const RegisterWrapper = styled.div`
+const LoginWrapper = styled.div`
   display: flex;
   justify-content: center;
   padding-top: 64px;
